[REF] core: separated odoo.orm.identifier
Define identifiers of records in their package.
