+++ odoo/api/__init__.py
@@ -1,9 +1,9 @@
 # ruff: noqa: F401
 # Exports features of the ORM to developers.
 # This is a `__init__.py` file to avoid merge conflicts on `odoo/api.py`.
+from odoo.orm.identifiers import NewId
 from odoo.orm.api import (
     Environment,
-    NewId,
     autovacuum,
     call_kw,
     constrains,
